requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
django-rest-knox==4.1.0
cryptography==3.4.7
black==20.8b1
mysqlclient==2.0.3
whitenoise==5.2.0
django-heroku==0.3.1